register half made

+++ db/migrate/20200513184334_create_users.rb
@@ -5,7 +5,7 @@ class CreateUsers < ActiveRecord::Migration[6.0]
       t.string :password_digest
       t.string :email
       t.date :dob
-      t.integer :steamID64
+      t.bigint :steamID64
       t.string :bio
 
       t.timestamps
